requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.8
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.51.0
grpcio==1.30.0
httplib2==0.18.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
moviepy==1.0.3
numpy==1.19.0
oauthlib==3.1.0
Pillow==7.1.2
pkg-resources==0.0.0
proglog==0.1.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.24.1
pytube3==9.6.4
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
srt==3.3.0
tqdm==4.46.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
